+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Minor: `F5` now reloads the badges
 - Minor: Homies global emotes can now be enabled/disabled
 - Minor: Added channel name & link to Twitch emotes popup
+- Minor: Added option to gray-out historical messages (Credits to Dankerino)
 
 - Minor: Ignore out of bounds check for tiling wms (#3270)